refactor(Repeater): Split VirtualizationInfo/AutomationPeer/TabFocusIterable per winui-port
Splits the per-element state and accessibility types per the winui-port file layout.
The duplicate ElementOwner.cs is removed and consolidated into VirtualizationInfo.h.mux.cs
to match the WinUI header. Bodies are 1:1 with WinUI commit 4b206bce3.

Tracks issue #22908.
